Add edge triggering for interrupt setup.

+++ src/interrupt_handlers.h
@@ -6,7 +6,7 @@
 #include "STM32L1xx.h"
 
 
-#define CLEAR_EXTI_PR(n) (0x1 << (n))
+#define CLEAR_EXTI_PR(n) ((uint32_t) (0x1 << (n)))
 
 
 /* User push button interrupt handler.
