+++ src/setup.c
@@ -29,9 +29,9 @@ void setup_interrupts()
     SYSCFG_EXTICR1 = (SYSCFG_EXTICR1 & ~EXTICRn_PA(0)) | EXTICRn_PA_EN(0); /* PA0. */
     SYSCFG_EXTICR1 = (SYSCFG_EXTICR1 & ~EXTICRn_PA(1)) | EXTICRn_PA_EN(1); /* PA1. */
 
-	/* Setup falling edge trigger for PA0 and PA1. */
-    EXTI->FTSR = (EXTI->FTSR & ~((uint32_t) EXTI_EDGE_EN(0))) | EXTI_EDGE_EN(0);
-    EXTI->FTSR = (EXTI->FTSR & ~((uint32_t) EXTI_EDGE_EN(1))) | EXTI_EDGE_EN(1);
+    /* Setup falling edge trigger for PA0 and PA1. */
+    EXTI->FTSR = (EXTI->FTSR & ~EXTI_EDGE_EN(0)) | EXTI_EDGE_EN(0); /* PA0. */
+    EXTI->FTSR = (EXTI->FTSR & ~EXTI_EDGE_EN(1)) | EXTI_EDGE_EN(1); /* PA1. */
     
     /* Unmask EXTI0 and EXTI1 interrupt using EXTI module. */
